address nate review feedback
