+++ lib/integration/github/account.ts
@@ -17,22 +17,20 @@
  */
 
 import { Octokit } from 'octokit';
-import { Endpoints } from '@octokit/types';
-import 'server-only';
 
-type ListRepositoriesResponse =
-  Endpoints['GET /user/installations/{installation_id}/repositories']['response']['data'];
+import 'server-only';
+import { SimplifiedRepo } from './types';
 
 /**
  * Fetches the repositories made available to a Github App installation by a
  * User Access Token. This function returns the full object provided by the
  * Github REST API, which should often be filtered as it contains far more
  * information than needed for any Tasktix use case.
- * @param accessToken
+ * @param accessToken Github Access Token generated by OAuth linking an account
  */
 export async function getAccessibleRepositories(
   accessToken: string
-): Promise<ListRepositoriesResponse['repositories'] | false> {
+): Promise<SimplifiedRepo[] | false> {
   const octokit = new Octokit({ auth: accessToken });
 
   try {
@@ -62,7 +60,17 @@ export async function getAccessibleRepositories(
       }
     );
 
-    return repos.data.repositories ?? false;
+    if (!repos.data.repositories) return false;
+
+    const simplifiedRepos: SimplifiedRepo[] = repos.data.repositories.map(
+      repo => ({
+        id: repo.id,
+        name: repo.full_name,
+        description: repo.description
+      })
+    );
+
+    return simplifiedRepos;
   } catch (error) {
     console.error('GitHub API Error:', error);
